chore: benchmark crypto service
